re-enable proxy cache in conda-python-tests jobs (#76)
In a previous PR, we'd temporarily turned off the proxy cache for conda
packages here, to work around some CI failures.

Details:
https://github.com/rapidsai/rapids-multi-gpu/pull/72#issuecomment-2660149987

Opening this to re-enable it. @ajschmidt8 made some configuration
changes (increasing the size of the buffer the proxy uses to hold
responses, to allow for larger repodata from conda channels), using this
to test if that worked.

---------

Co-authored-by: Peter Andreas Entschev <peter@entschev.com>