Do not bind CPUs in CI in `mpirun` and set custom container options (#93)
Running processes with `mpirun` will bind processes specified via `-np` to the CPU core with the same ID as the MPI process rank. This may be the source of instabilities, in particular if there are other jobs on the same host that happen to have high CPU load.

Additionally set custom container options that are important to provide reasonable resource limits for UCX, the same options are used in CI of both UCXX and UCX-Py.

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Mads R. B. Kristensen (https://github.com/madsbk)
  - Jake Awe (https://github.com/AyodeAwe)

URL: https://github.com/rapidsai/rapids-multi-gpu/pull/93