Avoid printing null-terminators (#102)
The PCI bus ID is retrieved by passing a pre-allocated string that will be overwritten but null-terminators will remain, when the `std::string` is redirected to a `std::stringstream` the null-terminators are treated as regular characters and printed as such. Those are most of the time invisible, but when the output is redirected to a file or another program such as `grep` it may be a identified as a binary file due to the presence of null-terminators. This change ensures only valid characters are redirected to the `std::stringstream`.

Additionally, replace the use of `"\n"` line breaks with the C++ `std::endl` instead.

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Mads R. B. Kristensen (https://github.com/madsbk)

URL: https://github.com/rapidsai/rapids-multi-gpu/pull/102