Static typing fixes (#115)
This updates our static typing configuration to disallow untyped definitions in `rapidsmp.*`. It fixes the issues that were violating that stricter rule.

(done properly this time, not pushing to main)

Probably the most controversial thing here is to type check everything, including tests. I'll usually forget the `-> None` from a test function and get annoyed, but type checking tests has caught enough of my mistakes in the past for it to be worth it to me. We kind of caught one here, where `format_bytes` was declared as accepting an `int` but we were using a float in `test_utils.py` (fixed by accepting `int | float` in `format_bytes`)

We'll need to synchronize this with some outstanding PRs, which might fail with the new, stricter rules. I'm happy to do that in either order.

Authors:
  - Tom Augspurger (https://github.com/TomAugspurger)

Approvers:
  - Mads R. B. Kristensen (https://github.com/madsbk)

URL: https://github.com/rapidsai/rapids-multi-gpu/pull/115