Update to RAPIDS 25.06 (#153)
This repository is still not using the same branching strategy of the rest of RAPIDS and probably also not setup for automatic updates. This change updates to RAPIDS 25.06.

Also remove the explicit `dask` dependency from `docs`, since that is picked transitively with `dask-cuda`.

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Jake Awe (https://github.com/AyodeAwe)

URL: https://github.com/rapidsai/rapids-multi-gpu/pull/153