Release `SpillManager` pointer without GIL (#208)
Release the GIL and reset C++ pointers in `SpillManager.__dealloc__()`. This is an attempt to resolve #207 , where the main thread has the GIL and is attempting to call `SpillManager::~SpillManager()` and join the thread it owns, while simultaneously the spill manager thread owned by the `SpillManager` object is attempting to acquire the GIL to run `SpillManager::spill()` given it registers Python callbacks.

With this change the GIL should be released when `SpillManager::~SpillManager()` is called and allow the `SpillManager` thread to acquire the GIL and complete running, and then the main thread without the GIL should be finally able to join the other thread.

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Richard (Rick) Zamora (https://github.com/rjzamora)

URL: https://github.com/rapidsai/rapids-multi-gpu/pull/208