Improve the performance of rapidsmpf_shuffle_graph
We've observed some slowness in
`rapidsmpf.integrations.dask.shuffler.rapidsmpf_shuffle_graph`, where a
relatively large amount of time (100s of ms) is spent in `client.run` staging things
for the shuffle.

This is especially pronounced when using the ucx and ucxx protocol and
as the number of workers increase. With Dask's tcp executor, the time
spent in `client.run` is on the order of 1-10 ms. With ucx that goes
to 100s of ms.

Long term, we want to address this at the root: our needs are relatively
simple here and we should be able to speed this up in dask / ucx-py /
ucxx.

Short term, we speed things up by only calling `client.run` once instead
of twice by combining the two things (getting worker ranks and staging
the shuffler) into the same function.