Use transition states in PausableThreadLoop
Since pausing and stopping a thread does not interrupt the execution
of the current iteration of its event loop, to correctly track state,
we introduce intermediate stopping and pausing states.

This is somewhat delicate because we were previously making (I think
incorrect) assumptions about the state of the thread loop after
calling pause. Fix those with some judicious wakes.

- Closes #362