Implement a single process communicator and make UCXX optional (#357)
Streaming execution on a single process might well want to use rapidsmpf for its spilling capabilities. In this scenario, we don't need the machinery of either MPI or UCXX, since all the data lives in a single process.

To facilitate this, build a "single-process" implementation of the `Communicator` interface. Since one is not expected to send to self, all of the communication routines just throw.

- Closes #326

Authors:
  - Lawrence Mitchell (https://github.com/wence-)
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Niranda Perera (https://github.com/nirandaperera)
  - Mads R. B. Kristensen (https://github.com/madsbk)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/rapidsmpf/pull/357