Store ready_ack_receives_ by rank to maintain order when completing
When we use UCXX as the communication backend, the test_some
implementation we have is not atomic with respect to background
progress. Although UCX matches tag messages in order, it might be that
when iterating over the requests we observe req[i] as uncompleted, and
then req[i+1] as completed (due to background progress). When this
occurs req[i] was also completed, but we didn't go back and check.
This breaks the message ordering responding to acks with the data
send. To avoid it, separate acks by rank (we don't care about ordering
between different ranks) and as soon as we observe a "gap" in the
completed requests, break out of the loop.