Remove `RegisterRank` control op from UCXX communicator (#408)
The `RegisterRank` control operation was initially added as an optimization, where a UCX endpoint handle (in the listener callback) would be associated to the rank before the UCXX endpoint was made available and properly registered. In practice that never occurs, `associate_endpoint_rank` only completes after `get_endpoint(rank, ucxx_endpoint)` is called on both processes that communicate with each other, and that would cause `rank_to_endpoint_`. Overwriting `rank_to_endpoint_` per se wouldn't be a problem, but there was also an error in the `RegisterRank` message where the remote rank would be sent to the remote endpoint (`rank`) instead of the local endpoint (`shared_resources_->rank()`).

The number of endpoints being created before and after changes contained here remain the same, in tests exactly `N^2` endpoints are created, for example running `4` ranks incurs in exactly `16` endpoints being created, therefore in practice this change is a no-op for existing use cases. The case where the existing implementation is problematic only happens when a rank communicates to self which was previously unused/unsupported.

Ultimately, this change also makes the implementation simpler and easier to understand.

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)
  - Niranda Perera (https://github.com/nirandaperera)

Approvers:
  - Niranda Perera (https://github.com/nirandaperera)

URL: https://github.com/rapidsai/rapidsmpf/pull/408