`PackedData` is memory-type agnostic (#404)
The gpu data of `PackedData` is now a `Buffer`, which can be either host or device memory.

This means that we don't have to _unspill_ gpu data before inserting it into the shuffler. But for now, `Shuffler::extract()` unspills the buffer so consumers like `unpack_and_concat()` don't have to provide a memory reservation.

The shuffler and the partitioning functions now take a `BufferResource` instead of a memory resource.

Authors:
  - Mads R. B. Kristensen (https://github.com/madsbk)

Approvers:
  - Lawrence Mitchell (https://github.com/wence-)
  - Niranda Perera (https://github.com/nirandaperera)

URL: https://github.com/rapidsai/rapidsmpf/pull/404