Remove `mem_type` from some ops in Buffer API (#410)
In the `BufferResource` utils such as `copy`, we accept a `mem_type` and a `reservation`. I think this is redundant and error prone. 
One could provide a conflicting mem_type and reservation, device/ host - We throw when this happens. But IMO this is unnecessary, because allocation is already made for the required mem_type. 
Also we still have the inconsistency reported in #280. 

So, this PR removes the redundant mem_type, and forces buffer utils to use reservations to make new allocations and removes buffer resource/ mem type args. This also this fixes #280

~~Also this PR makes buffer resource `move*` methods static, as `Buffer` no longer requires `br*`~~

Authors:
  - Niranda Perera (https://github.com/nirandaperera)

Approvers:
  - Lawrence Mitchell (https://github.com/wence-)

URL: https://github.com/rapidsai/rapidsmpf/pull/410