Fixed exception handling in Statistics.get_stat
This fixes the segfault during `Statistics.get_stat`. Previously, these
methods were marked as `nogil`, but we need to catch the C++
`std::out_of_range` so now they are `except +`.

Now we'll raise a `KeyError` instead of segfaulting.

Also fixes an issue with the type on `cpp_get_statistic_value`, which
should be a double rather than a size_t.

Closes https://github.com/rapidsai/rapidsmpf/issues/421