Fix matching of completed futures to chunks in allgather (#474)
The lack of message overtaking in MPI and UCX means that messages are matched in order to their receive buffers. However this need not mean that the set of completed futures when we call test_some is from a contiguous block (even if all messages are from the same rank with the same tag). In the allgather implementation we were incorrectly relying on this (incorrect) invariant.

To fix it, return the indices of completed futures in the input future vector along with the completed futures themselves. This allows us to match up completed futures with their correct chunk to reattach data buffers.

- Closes #469
- Closes #472

Authors:
  - Lawrence Mitchell (https://github.com/wence-)

Approvers:
  - Mads R. B. Kristensen (https://github.com/madsbk)

URL: https://github.com/rapidsai/rapidsmpf/pull/474