Remove possibility of race condition when using extract_ready
The AllGather implementation offers a non-blocking way of extracting
any ready messages in an unordered fashion by allowing us to extract
whichever chunks have currently been received.

However, the implementation previously admitted a race condition where
an extracting thread could incorrectly observe that the allgather
operation was finished and thus exit its extraction loop before all
chunks have been received.

The culprit is a lack of ordering between the modification of the
finish counter and moving the extraction postbox's goalpost. To fix
this, ensure that when we receive a finish chunk we first increment
the goalpost and then decrement the finish counter and introduce a
happens-before relationship between the two.

- Closes #480