Introduce a throttled channel object
This channel behaves identically to a standard `Channel` except that
one must acquire tickets granting permission to send before one is
able to send. The send then returns a receipt that grants permission
to release the ticket.

This pattern is very useful for producer nodes where we want some
amount of bounded concurrency in the tasks that might suspend before
sending into a channel.