Fix hang in boostrap_dask_cluster with new clients
The resources created by rapidsmpf are all global to the scheduler /
worker processes.

Dask supports multiple Clients sharing the same cluster.

Previously, we used the Client ID to track whether we've bootstrapped
a Dask Cluster. If the user provided a new Client for the same cluster,
we'd attempt to re-bootstrap it and things would hang.

To fix this, we use the scheduler ID.