Add optional owner slot to `TableChunk`s (#536)
When a `TableChunk` is constructed from a `cudf::table_view`, we are on the hook to keep the backing data alive. When we move `TableChunk`s into `Message`s and out again from Python we cannot do this in a sane way externally, so we must stash the owning object in the `TableChunk` itself.

The `PyObject` is stored in a `unique_ptr` with a custom deleter that, when called, acquires the GIL and decrefs the object. This way, even if we consume the `TableChunk` in a C++ node (rather than a Python node) its backing storage will be deallocated.

Authors:
  - Lawrence Mitchell (https://github.com/wence-)

Approvers:
  - Mads R. B. Kristensen (https://github.com/madsbk)
  - Gil Forsyth (https://github.com/gforsyth)

URL: https://github.com/rapidsai/rapidsmpf/pull/536