Ensure messages are dropped by refcount through channel sends (#638)
The previous approach of defining local functions seems like it captures objects in a refcycle that is then only broken by the cyclic collector. This means that if we just sink messages without releasing their payload they might not get collected until the interpreter runs atexit teardown. If the destruction needs to deallocate device memory, the memory resource will already be gone leading to a segfault.

To fix this, use functools.partial to make the closures we pass to the C++ callbacks.

Authors:
  - Lawrence Mitchell (https://github.com/wence-)

Approvers:
  - Peter Andreas Entschev (https://github.com/pentschev)
  - Mads R. B. Kristensen (https://github.com/madsbk)

URL: https://github.com/rapidsai/rapidsmpf/pull/638