Separate shuffler's communication into new interface (#437)
Move communication operations from the `Shuffler::ProgressThread` into a new `CommunicationInterface` abstract class to handle all communication, with the use of a class `Message`. This approach has the benefit of making a more distinct and easier to understand separation of the communication routine from the shuffler making the progress operator considerably smaller (will be done in follow-up PR), simultaneously allowing extensions to the communication interface disjoint from the shuffler, for example, to allow implementing an Active Message-based communicator for the shuffler. It also generalizes the concept of a "message", instead of transferring chunks all the `CommunicationInterface` knows about are messages represented by the `Message` class that combines metadata, payload and src/dst rank.

Currently implement only a `TagCommunicationInterface` that implements exactly the behavior that is implemented in the shuffler progress, additionally removing the need for the ack message. Extending to support Active Messages will also be handled in a separate PR.

Removes also ack messages, closes #475, closes #535 .

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Mads R. B. Kristensen (https://github.com/madsbk)
  - Lawrence Mitchell (https://github.com/wence-)

URL: https://github.com/rapidsai/rapidsmpf/pull/437