Implement linearisation by round-robin assignment (#654)
Without buffering in the `Lineariser`, with random task assignment to producers, we must serialise pulling data out of the producers since when we empty a slot we might need to refill from the same slot to get the next sequence number.

To fix this, and get decent pipelined throughput, use round-robin assignment of tasks to producers in the lineariser. To handle the buffering we now use a queue per producer whose buffer size is controlled by a per-queue semaphore. Producers acquire a ticket to send, similar to the `ThrottlingAdaptor`, but the release is handled by the consumer: receiving from a `BoundedQueue` gives us the message and a receipt to release once we have handled the message.

At the same time, improve the `read_parquet` implementation to avoid reading too much metadata too many times.

Authors:
  - Lawrence Mitchell (https://github.com/wence-)

Approvers:
  - Mads R. B. Kristensen (https://github.com/madsbk)

URL: https://github.com/rapidsai/rapidsmpf/pull/654