Fix multi-rank parquet read bug
If a rank had no files to read it would deduce it had no producers and
hence try and send an empty table of the correct shape. But in doing so it
would try and access its local_files array, which is empty.

Instead only send an empty table if we have at least one local file.