Fix retrieving number of ranks when running benchmarks with `rrun` (#709)
Retrieving the number of ranks is written as part of `ArgumentParser` in the C++ benchmarks that occurs before initializing the communicator. Because of that, `rrun` was defaulting to only one rank in internals which led to wrongly calculating global throughput as if it was just the local throughput (always multiplying by only one rank).

This change adds a new helper function to retrieve the number of ranks from `rrun` before initializing the communicator to mimic what's done with MPI and ensure global throughput is correctly calculated.

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Tom Augspurger (https://github.com/TomAugspurger)

URL: https://github.com/rapidsai/rapidsmpf/pull/709