Pin Ray<2.52 (#730)
Some Ray tests have been non-deterministically segfaulting, the time the segfaults started coincided with the date when `ray=2.52` was release, therefore there's reason to believe it's a bug in that release. For now, pin `ray<2.52` to determine if that is in fact the source of the problem.

See https://github.com/rapidsai/rapidsmpf/pull/702 for previous discussion.

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Mads R. B. Kristensen (https://github.com/madsbk)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/rapidsmpf/pull/730