Fix `non_device_size` calculation in unpack_and_concat (#765)
I believe this is intended to reflect the number of bytes not in device memory. We'll want to examine the other usage of `non_device_size` (in `br->reserve_device_memory_and_spill`) to ensure that we're OK with this change. IIUC, we'll end up reserving additional bytes equal to the number of bytes that were not in GPU memory.

Authors:
  - Tom Augspurger (https://github.com/TomAugspurger)

Approvers:
  - Niranda Perera (https://github.com/nirandaperera)

URL: https://github.com/rapidsai/rapidsmpf/pull/765